% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findIsotopes.R
\name{calcIsotopeMatrix}
\alias{calcIsotopeMatrix}
\title{Create a matrix for isotope abundanc computation and id}
\usage{
calcIsotopeMatrix(maxiso = 4)
}
\arguments{
\item{maxiso}{max number of isotopes}
}
\value{
matrix
}
\description{
Create a matrix for isotope abundanc computation and id
}
