% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_global_opt.R
\name{checkDependencies}
\alias{checkDependencies}
\title{Check for installed dependencies}
\usage{
checkDependencies()
}
\description{
Checks for installed dependencies and then prompts the user to install if needed.
Called by readGlobalOptions().
}
