% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_formula.R
\name{compute_formulae}
\alias{compute_formulae}
\title{Compute molecular formulae for a list of masses in the database}
\usage{
compute_formulae(DBE = 5)
}
\arguments{
\item{DBE}{the minimum required double bond equivalence}
}
\description{
Compute molecular formulae for a list of masses in the database
}
