% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_convertRAW.R
\name{convertRaw}
\alias{convertRaw}
\title{Convert raw data}
\usage{
convertRaw(raw_file_dir, msconvert_path)
}
\description{
convertRaw converts vendor raw data using msconvert into an mzML file. This mzML file is then used by segmentmzML data to segment the file into SIM m/z ranges.
}
\examples{
#convertRaw()


}
