% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_peak_groups.R
\name{count_peak_groups}
\alias{count_peak_groups}
\title{count_peak_groups}
\usage{
count_peak_groups(mzmin, mzmax, mzwid)
}
\arguments{
\item{mzmin}{the minimum mz value}

\item{mzmax}{the maximum mz value}

\item{mzwid}{the mz slice \(default = 0.1\)}
}
\description{
count_peak_groups
}
