% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hit_table.R
\name{create_hit_table}
\alias{create_hit_table}
\title{Create a hit table of results in the SQLite data base.}
\usage{
create_hit_table()
}
\description{
Create a hit table of results in the SQLite data base.
}
