% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ref_table.R
\name{create_ref_table}
\alias{create_ref_table}
\title{Creates a reference  / master table of compounds for RT alingments}
\usage{
create_ref_table(table_name, db.list, ppm, rt_tol)
}
\arguments{
\item{table_name}{the name of the SQLlite DB to output.   Do not add .sqlite to the name}

\item{db.list}{the list of SQLlite DBs to combine}

\item{ppm}{the ppm tolerance for matching the masses}

\item{rt_tol}{the rt toleracne in minumtes for searching for a mass peak in the ref DB}
}
\description{
Creates a reference  / master table of compounds for RT alingments
}
