% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formula_counts.R
\name{delta_formula}
\alias{delta_formula}
\title{Compute the delta formula for two features, the precursor and the aglycone}
\usage{
delta_formula(f1, f2)
}
\arguments{
\item{f1}{the first formula}

\item{f2}{the aglycone formula}
}
\value{
numeric delta value
}
\description{
Compute the delta formula for two features, the precursor and the aglycone
}
