% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{detectPeak}
\alias{detectPeak}
\title{Detect a peak in an XIC}
\usage{
detectPeak(XIC, rt)
}
\arguments{
\item{XIC}{matrix of the XIC}

\item{rt}{numeric rt of the peak in minutes}
}
\value{
matrix of the detected peak
}
\description{
Detect a peak in an XIC
}
