% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{detectPeakAtRT}
\alias{detectPeakAtRT}
\title{Detect peaks in an XIC}
\usage{
detectPeakAtRT(XIC, rt)
}
\arguments{
\item{XIC}{matrix XIC of 3 columns (zmz, rt, intensity)}

\item{rt}{numeric target retention time}
}
\value{
list
}
\description{
Detect peaks in an XIC
}
