% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlaplace}
\alias{dlaplace}
\title{Probability density function: Laplace}
\usage{
dlaplace(X, m, b)
}
\arguments{
\item{X}{numeric vector of values}

\item{m}{numeric the mode of the distribution}

\item{b}{numeric scale parameter}
}
\value{
vector of numeric values containing density values
}
\description{
Generates the tolerance window in Da for a given ppm value and mass
}
\examples{
x <- seq(-10,10,by = 0.001)
d_lap <- dlaplace(x,m=0,b = 1)
plot(x,d_lap, type = "l")
}
