% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_known_adducts.R
\name{find_known_adducts}
\alias{find_known_adducts}
\title{Find known adducts by searching against a database of known adducts. Usually spike ins, but a knwons db is included.}
\usage{
find_known_adducts(search_tol = 10, minscore = 0.5)
}
\arguments{
\item{search_tol}{the mass search tol (ppm).}

\item{minscore}{the minimum match score for the match to a known.}
}
\value{
dataframe of matches.
}
\description{
Find known adducts by searching against a database of known adducts. Usually spike ins, but a knwons db is included.
}
