% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{find_peaks}
\alias{find_peaks}
\title{Find peaks from a vector of intensities}
\usage{
find_peaks(x, m = 3)
}
\arguments{
\item{x}{vector of intensities}

\item{m}{minimum distance between peaks}
}
\value{
vector of peak indices
}
\description{
Find peaks from a vector of intensities
}
