% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findIsotopes.R
\name{findIsotopes}
\alias{findIsotopes}
\title{Find isotopic peaks in peak data}
\usage{
findIsotopes(mslevel = 1, maxiso = 3)
}
\arguments{
\item{mslevel}{the ms level to search}

\item{maxiso}{the maximum number of isotopes}
}
\description{
Find isotopic peaks in peak data
}
