% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findIsotopesPspec.R
\name{findIsotopesPspec}
\alias{findIsotopesPspec}
\title{Find isotopes in psuedo spectra}
\usage{
findIsotopesPspec(
  isomatrix,
  dat,
  ppmtol = 15,
  maxiso = 4,
  maxcharge = 2,
  mzabs = 0.005,
  minfrac = 0.5
)
}
\arguments{
\item{isomatrix}{the isomatrix}

\item{dat}{the data frame}

\item{ppmtol}{the ppm tolerance}

\item{maxiso}{max # isotopologues}

\item{maxcharge}{max allowable charge}

\item{mzabs}{max mz deviation in  the search}

\item{minfrac}{min sample fraction (not used)}
}
\value{
matrix
}
\description{
Find isotopes in psuedo spectra
}
