% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_feature.R
\name{findPeaksMSDIAL}
\alias{findPeaksMSDIAL}
\title{Find features using MSDIAL}
\usage{
findPeaksMSDIAL(sample_directory, scandef_file, msdial_path,
  msdial_param_path)
}
\arguments{
\item{sample.result.path}{the path to the directory where the results will be stored, read using the global parameters file.}
}
\description{
A function which calls the MSDIAL Console application.  The location of the console appplication must be provided in the
global parameters file.   There must also be a msdial parameters file included.
}
\examples{

 #findPeaksMSDIAL("results/sample1/")

}
