% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_feature.R
\name{findPeaksMSDIALSample}
\alias{findPeaksMSDIALSample}
\title{Find peaks using MSDIAL, multiple samples}
\usage{
findPeaksMSDIALSample(sample_directories, scandef_file, msdial_path,
  msdial_param_path)
}
\description{
Wrapper function to run MSDIAL on multiple samples.
}
\examples{

findPeaksMSDIAL_sample()

}
