% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formula_counts.R
\name{get_atom_counts}
\alias{get_atom_counts}
\title{Retrieve the atomic counts and mass defect for a given formula / mz}
\usage{
get_atom_counts(f, mz = NULL, kendrik_formula = NULL)
}
\arguments{
\item{f}{the molecular formula}

\item{mz}{the exact mass}

\item{kendrik_formula}{character vector string of the molecular formulae}
}
\value{
data.frame of results
}
\description{
Retrieve the atomic counts and mass defect for a given formula / mz
}
