% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alignment_data.R
\name{getAlignmentData}
\alias{getAlignmentData}
\title{Get alignment data from aligned DNA adducts}
\usage{
getAlignmentData(db_name)
}
\arguments{
\item{db_name}{name of the SQL table}
}
\value{
list of all tables (data frames) in the SQL database
}
\description{
Get alignment data from aligned DNA adducts
}
