% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_IS.R
\name{getInternalStandards}
\alias{getInternalStandards}
\title{getInternalStandards}
\usage{
getInternalStandards(msdial_results)
}
\arguments{
\item{list}{containing msdial.result}
}
\description{
Finds any internal standards that might exist in the search results and indexes them.
}
\examples{
getInternalStandards()
}
