% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getMassTolRange}
\alias{getMassTolRange}
\title{Get mass tolerance range in ppm}
\usage{
getMassTolRange(m, ppm)
}
\arguments{
\item{m}{a double number, the mass}

\item{ppm}{a double number, the user provided ppm tol window}
}
\value{
vector of numeric values containing min and max range in Da
}
\description{
Generates the tolerance window in Da for a given ppm value and mass
}
\examples{
getMassTolRange(m = 490.1946,ppm = 5)
490.1921 490.1971
}
