% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_processFeatures.R
\name{getMSDIAL_results}
\alias{getMSDIAL_results}
\title{getMSDIAL_results}
\usage{
getMSDIAL_results(sample_directories)
}
\arguments{
\item{sample_results_path}{character vector containing the parent directory of results}
}
\value{
list of data.frames containing wsim and nl feature data produced using MSDIAL
}
\description{
getMSDIAL_results is a wrapper function for running msdial_combineDataFiles.
}
\examples{

multipleResults <- getMSDIAL_results("results")


}
