% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_nlm.R
\name{getNLMScore}
\alias{getNLMScore}
\title{The the NLM score}
\usage{
getNLMScore(searchResultList, mod_mz)
}
\arguments{
\item{searchResultList}{}

\item{mod_mz}{}
}
\value{
list of search result data frames
}
\description{
Computes the NL model score for each dM value.
}
