% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{getPeak}
\alias{getPeak}
\title{Get the peak}
\usage{
getPeak(spectrum, mz, ppm)
}
\arguments{
\item{spectrum}{matrix of mz and intensities}

\item{mz}{target mz value}

\item{ppm}{tolerance window in ppm}
}
\value{
list of mz and intensities
}
\description{
Get the peak
}
