% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{getPeaks}
\alias{getPeaks}
\title{Get the peaks from an XIC pair}
\usage{
getPeaks(ms1_XIC, ms2_XIC, adduct_mass, rt, sp = 0.4, plot = TRUE,
  filename = NULL)
}
\arguments{
\item{ms1_XIC}{XIC matrix from the ms1 data}

\item{ms2_XIC}{XIC matrix from the ms2 data}

\item{adduct_mass}{the adduct mz value.  Dont forget the "-" if a loss.}

\item{rt}{numeric retention time value}

\item{sp}{numeric span for the smoother}

\item{plot}{boolean plot or no plot}
}
\value{
dataframe
}
\description{
Get the peaks from an XIC pair
}
