% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{getPlots}
\alias{getPlots}
\title{getPlots}
\usage{
getPlots(searchResultList, scandef_file, sampleDir, min_score = 0.95,
  min_intensity = 1000, ppm = 6, rt_tol = 1.5, min_ratio = 10)
}
\arguments{
\item{searchResultList}{}

\item{scandef_file}{}

\item{sampleDir}{}

\item{min_score}{}

\item{min_intensity}{}

\item{ppm}{}

\item{rt_tol}{}
}
\description{
getPlots
}
