% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_convertRAW.R
\name{getRawFileList}
\alias{getRawFileList}
\title{Get the list of raw files}
\usage{
getRawFileList(raw_file_dir)
}
\arguments{
\item{character}{vector raw_file_dir}
}
\value{
vector of raw file names
}
\description{
Returns a list of .raw files.
}
