% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{getRawXIC}
\alias{getRawXIC}
\title{Extract the XIC from raw data}
\usage{
getRawXIC(spectra, header, mz, ppm = NULL, mzdec = NULL, rtmin, rtmax,
  smooth = FALSE, sp = NULL)
}
\arguments{
\item{spectra}{object mzR spectra}

\item{header}{object msR header infor}

\item{mz}{numeric desired mz to extract}

\item{ppm}{numeric tolerance window in ppm (7 is standard in our system)}

\item{mzdec}{}

\item{rtmin}{numeric minimum rt}

\item{rtmax}{numeric maximum rt}

\item{smooth}{boolean}

\item{sp}{numeric span value}
}
\value{
matrix
}
\description{
Extract the XIC from raw data
}
