% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{getRTRangeByRT}
\alias{getRTRangeByRT}
\title{getRTRangebyRT}
\usage{
getRTRangeByRT(RT, tol)
}
\arguments{
\item{RT}{numeric retention time}

\item{tol}{numeric tolerance}
}
\value{
vector containing limits of RT range
}
\description{
getRTRangebyRT
}
