% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_convertRAW.R
\name{getSampleDirectories}
\alias{getSampleDirectories}
\title{Get sample directories}
\usage{
getSampleDirectories(results_path)
}
\arguments{
\item{results_path}{character vector indicating the directory in which to search.}
}
\value{
vector of sample directories
}
\description{
Get sample directories
}
