% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_convertRAW.R
\name{getSampleNames}
\alias{getSampleNames}
\title{Get the sample names}
\usage{
getSampleNames(raw_file_dir)
}
\arguments{
\item{raw_file_dir}{character vector indicating the raw file directory to search.}
}
\value{
vector of sample names
}
\description{
Gets the sample names indicated using the raw file names.
}
