% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{getSIMWindows}
\alias{getSIMWindows}
\title{Get the list of sim windows}
\usage{
getSIMWindows()
}
\value{
a character vector of scan windows
}
\description{
Retrieves the scan windows for plotting functions.
}
