% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_formula.R
\name{getValid}
\alias{getValid}
\title{get valid atomic formulae}
\usage{
getValid(flist, DBE = 4, precursor = T)
}
\arguments{
\item{flist}{list of molecular formulae}

\item{DBE}{minimum double bond equivalent}

\item{precursor}{is a precursor}
}
\value{
list containing: forula, DBE, exact_mass
}
\description{
get valid atomic formulae
}
