% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{getXIC}
\alias{getXIC}
\title{Get an extracted ion chromatogram (XIC)}
\usage{
getXIC(spectra, header, mz, ppm = NULL, mzdec = NULL, rt, rt_tol,
  rt_offset = NULL, smooth = TRUE, sp = NULL)
}
\arguments{
\item{spectra}{list of mzR spectra objects}

\item{header}{header extracted from mzML using mzR}

\item{mz}{numeric containing the mz value}

\item{ppm}{numeric the ppm tolerance window to extract the mz value}

\item{rt}{numeric containing the retention time in seconds}

\item{rt_tol}{numeric containing the retention time range to extract the XIC}
}
\value{
matrix of smoothed retention times (min) and intensities
}
\description{
Helper function to extract an extracted ion chromatogram using mzR.
}
