% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{getXIC_cor}
\alias{getXIC_cor}
\title{Extract XIC data and compute correlations on peaks.}
\usage{
getXIC_cor(sampleDir, scandef_file, plot = TRUE)
}
\arguments{
\item{sampleDir}{character vector pointing to the sample directory containing the peak list.}

\item{scandef_file}{character vector pointing to the scan definition file.}
}
\value{
list of dataframes containing the correlation and scores for precursor NL pairs.
}
\description{
Extract XIC data and compute correlations on peaks.
}
