% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupFWHM.R
\name{groupFWHM}
\alias{groupFWHM}
\title{Group peaks by retention time}
\usage{
groupFWHM(sigma = 6, perfwhm = 0.6)
}
\arguments{
\item{sigma}{number of standard deviation arround the mean (6 = 2 x 3 left and right)}

\item{perfwhm}{percentage of width of peak at fwhm}
}
\description{
Group peaks by retention time
}
