% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_align.R
\name{join_align}
\alias{join_align}
\title{Perform retention time aligment using the joing align algorithm}
\usage{
join_align(ref_db, db.list, ppm = 7, rt_tol = 0.5, N)
}
\arguments{
\item{ref_db}{the SQLlite reference database}

\item{db.list}{the list of SQLlite DBs to align}

\item{ppm}{mz tolerance window in ppm}

\item{rt_tol}{rt tolerance window in minutes}

\item{N}{number of samples to align}
}
\description{
Perform retention time aligment using the joing align algorithm
}
