% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_other.R
\name{list.dirs.depth.n}
\alias{list.dirs.depth.n}
\title{List directories in folder}
\usage{
list.dirs.depth.n(p, n)
}
\arguments{
\item{p}{string path of directory.  Default = '.'}

\item{n}{integer value indicating the folder depth to search}
}
\value{
vector of strings containing the paths to the directories
}
\description{
List directories in folder
}
