% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_nlm.R
\name{load_search_object}
\alias{load_search_object}
\title{Load a list of search objects.}
\usage{
load_search_object(row_data)
}
\arguments{
\item{row_data}{vector row from data.frame}
}
\value{
list of class search_result
}
\description{
Load a list of search objects.
}
