% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_convertRAW.R
\name{makeSampleDir}
\alias{makeSampleDir}
\title{Make directories for each sample}
\usage{
makeSampleDir(result_dir, sample_names, scandef_file)
}
\description{
Creates directories for sample of interest indicateg in the global options parameter file.
}
\examples{

#makeSampleDir()
}
