% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_convertRAW.R
\name{makeSIMDir}
\alias{makeSIMDir}
\title{Create SIM directories}
\usage{
makeSIMDir(scandef_file, out_path)
}
\arguments{
\item{scandef_file}{a non empty character vector giving the wide SIM scan definition file.}

\item{out_path}{a non empty character vector giving the location used to write results.}
}
\description{
makeSIMDir creates sim directories for performing independent  MSDIAL results files.
}
\examples{
#makeSIMDir(out_path = "results")
}
