% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mdnadb}
\alias{mdnadb}
\title{Known modified DNA nucleotides.}
\format{
A data frame with 349 rows and 7 variables:
\describe{
  \item{Name}{common adduct name}
  \item{Formula}{Lewis formula}
  \item{DBID}{Modified DNA DataBase ID / accession}
  \item{PCID}{PubChem ID}
  \item{Exact.Mass}{the exact compound mass}
  \item{Precursor}{the precursor mz}
  \item{Aglycone}{the mz of the compound after removal of the ribose}
  ...
}
}
\usage{
mdnadb
}
\description{
A dataset containing the precursor masses of several known DNA adducts.
}
\keyword{datasets}
