% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_io.R
\name{merge_results}
\alias{merge_results}
\title{Merge MSDIal results}
\usage{
merge_results(sample_directories, adduct_list, score_feature = 0.9,
  totalScore = 0.8, maxRatio = 5, minIntensity = 3000)
}
\arguments{
\item{sample_directories}{vector of sample directory paths}

\item{score_feature}{minimum score}

\item{totalScore}{minimum total score}

\item{maxRatio}{maximum ratio}

\item{minIntensity}{minimum intensity}
}
\description{
Merge MSDIal results
}
