% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_io.R
\name{merge_searchResults}
\alias{merge_searchResults}
\title{Merge search results by adducts}
\usage{
merge_searchResults(sampleDir, adduct_list, score_feature, totalScore,
  maxRatio, minIntensity)
}
\arguments{
\item{sampleDir}{character vector path to the sample directory}

\item{adduct_list}{data.frame of adducts}

\item{score_feature}{numeric minimmum score feature}

\item{totalScore}{numeric mumimum total score}

\item{maxRatio}{numeric maximum ratio}

\item{minIntensity}{numeric minimum ratio}
}
\description{
Merges search results from the NLM search
}
