% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_tables.R
\name{merge_tables}
\alias{merge_tables}
\title{Merge two sqlite DB tables together for alignment}
\usage{
merge_tables(X, Y, ppm, rt_tol)
}
\arguments{
\item{X}{the first data frame}

\item{Y}{the second data frame}

\item{ppm}{the mz tolerance for searching for a mass in the ref DB}

\item{rt_tol}{the rt tolerance for searching for a mass peak in the ref DB in minutes}
}
\value{
data frame of merged tables
}
\description{
Merge two sqlite DB tables together for alignment
}
