% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_nlm.R
\name{modelNLM}
\alias{modelNLM}
\title{modelNLM}
\usage{
modelNLM(sampleDir, data_X, data_Y = NULL, adduct_list, boost = 2,
  alpha_mz = 0.5, beta_rt = 0.5, ppm_window = 30, rt_tol = 0.2,
  instrument_tol = 0.01)
}
\arguments{
\item{sampleDir}{the path to the sample to by analyzed}

\item{data_X}{data frame of MS1 features and associated data.}

\item{data_Y}{data frame of MS1 or MS2 and associated data.}

\item{adduct_list}{data.frame of adducts}

\item{boost}{numeric value indicating the boost factor for the weigths to be applied to the dM values (default = 3).}

\item{alpha_mz}{numeric value, 0-1 to weight the mz search score.}

\item{beta_rt}{numeric value, 0-1 to weight the rt search score. Alpha and beta must add up to 1.}

\item{ppm_window}{numeric value for the ppm search window, usually 30-50 ppm.}

\item{rt_tol}{numeric value for the retention time search window.}

\item{instrument_tol}{the instrument tolerance window for the search score, usually 10ppm.}
}
\description{
modelNLM
}
