% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_nlm.R
\name{modelNLM_run}
\alias{modelNLM_run}
\title{Run NLM model on multiple samples}
\usage{
modelNLM_run(msdial_results, instrument_tol = 5, boost = 3)
}
\arguments{
\item{msdial_results}{list of feature finding results from MSDIAL}

\item{instrument_tol}{number for the extpected  mass accuracy of the instrument in ppm}

\item{boost}{numeric 0-5 value for weighting the Laplace distribution}
}
\value{
list of nlm model results.
}
\description{
Wrapper function to run modelNLM for a list of samples.
}
\examples{

modelNLM_run(msdial_results)

}
