% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_processFeatures.R
\name{msdial_combineDataFiles}
\alias{msdial_combineDataFiles}
\title{Combine MSDIAL results}
\usage{
msdial_combineDataFiles(sample_directory)
}
\arguments{
\item{sample_result_path}{}
}
\value{
list containing two dataframes, one each for the MS1 and MS2 data.
}
\description{
Combines msdial data from the various SIM windows.
}
\examples{

#msdial_combineDataFiles("results/sample1/")

}
