% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_processFeatures.R
\name{msdial_data}
\alias{msdial_data}
\title{Read MSDIAL result file}
\usage{
msdial_data(fh = NULL)
}
\arguments{
\item{fh}{string}
}
\value{
data.frame
}
\description{
Reads a single .msdial data file and returns a data.frame
}
\examples{



#data <- msdial_data(fh = "results.msdial")

}
