% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_candidate.R
\name{plot_candidate}
\alias{plot_candidate}
\title{Plot candidate DNA adduct}
\usage{
plot_candidate(ppm = 7)
}
\arguments{
\item{ppm}{tolerance window in ppm}
}
\description{
Plot candidate DNA adduct
}
