% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dna_adduct_map.R
\name{plot_dna_adduct_map}
\alias{plot_dna_adduct_map}
\title{Plot a DNA adduct heat map}
\usage{
plot_dna_adduct_map(knowns = F)
}
\arguments{
\item{knowns}{boolean T / F plot only identified known DNA adducts?}
}
\value{
ggplot plot
}
\description{
Plot a DNA adduct heat map
}
