% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_peak_group.R
\name{plot_peak_group}
\alias{plot_peak_group}
\title{plot peaks that are DNA adducts.}
\usage{
plot_peak_group(pk_group = NULL, mz = NULL, bw = 0.35)
}
\arguments{
\item{pk_group}{numeric index}

\item{mz}{mz null}

\item{bw}{bandwidth}
}
\description{
plot peaks that are DNA adducts.
}
