% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plotKnowns}
\alias{plotKnowns}
\title{Plot XIC of known ions}
\usage{
plotKnowns(scandef_file, sampleDir, ppm = 6, rt_tol = 1.5)
}
\arguments{
\item{scandef_file}{}

\item{sampleDir}{}

\item{ppm}{}

\item{rt_tol}{}
}
\description{
Generates an EIC of a known compound.
}
