% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plotPeak.plotly}
\alias{plotPeak.plotly}
\title{Plots the spectrum of precursor and aglycones using plotly}
\usage{
plotPeak.plotly(spectrum, mz, ppm, main = NULL, col = 1, title = NULL)
}
\arguments{
\item{spectrum}{matrix of mz and intensities}

\item{mz}{target mz value}

\item{ppm}{tolerance window in ppm}

\item{main}{character vector for the plot title}

\item{col}{integer color value}

\item{titl}{character the title of the plot}
}
\description{
Plots the spectrum of precursor and aglycones using plotly
}
