% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plotPeak}
\alias{plotPeak}
\title{plots and XIC and detects the peak in it}
\usage{
plotPeak(ms1_mz, ms2_mz, ppm, rt, rt_tol, smooth = FALSE, sp = NULL,
  sampleDir = NULL, file = NULL)
}
\arguments{
\item{ms1_mz}{numeric ms1 mz}

\item{ms2_mz}{numeric ms2 mz}

\item{ppm}{numeric ppm tolerance for extraciton}

\item{rt}{numeric retention time in minutes}

\item{rt_tol}{numeric retention time window in minutes}

\item{smooth}{boolean T/F to apply a smoothing function}

\item{sp}{numeric the spar for the smoother, typically 0.12}

\item{sampleDir}{character vector of the sample directory containing the mzML files}

\item{title}{character vector of the title to use in the plot}
}
\description{
plots and XIC and detects the peak in it
}
