% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plotRAWPeak}
\alias{plotRAWPeak}
\title{Plot peaks from raw data}
\usage{
plotRAWPeak(ms1_mz, ms2_mz, ppm, rtmin, rtmax, smooth = FALSE,
  sp = NULL, sampleDir = NULL, file = NULL)
}
\arguments{
\item{ms1_mz}{}

\item{ms2_mz}{}

\item{ppm}{}

\item{rtmin}{}

\item{rtmax}{}

\item{smooth}{}

\item{sp}{}

\item{sampleDir}{}

\item{file}{}
}
\description{
Plot peaks from raw data
}
