% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ppmErr}
\alias{ppmErr}
\title{Calculate ppm error for a given mass}
\usage{
ppmErr(m, m0)
}
\arguments{
\item{m}{a double number, the experimental mass}

\item{m0}{a double number, the theoretical mass}
}
\value{
double number the ppm error
}
\description{
This function compares 2 masses and computes the ppm error between them,
}
\examples{
ppmErr(490.1944,490.1946)
 -0.4080012
}
