% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_io.R
\name{read_searchResultFile}
\alias{read_searchResultFile}
\title{read_searchResultFile}
\usage{
read_searchResultFile(fh, score_feature, totalScore, maxRatio,
  minIntensity)
}
\arguments{
\item{fh}{character vector file path}

\item{score_feature}{numeric minimum score feature}

\item{totalScore}{numeric minimum total score}

\item{maxRatio}{numeric max ratio}

\item{minIntensity}{numeric minimum intensity}
}
\value{
data.frame
}
\description{
Reads an msdial search result file
}
