% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_global_opt.R
\name{readGlobalOpts}
\alias{readGlobalOpts}
\title{Read the global options file and set global values}
\usage{
readGlobalOpts(fh)
}
\arguments{
\item{fh}{character vector indicating the path to the global options file provided by the user.}
}
\description{
Reads the global options parameter file and loads the variables into the global environment.
}
\examples{

readGlobalOpts("global_params.txt")
}
