% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_table.R
\name{refine_table}
\alias{refine_table}
\title{Rine table and remove duplicates}
\usage{
refine_table(merged_table, ppm = 10, rt_tol = 1.5)
}
\arguments{
\item{merged_table}{the original master table}

\item{ppm}{the mz tolerance window in ppm}

\item{rt_tol}{the retention time window in minutes}
}
\value{
data frame of the refined merged table
}
\description{
Rine table and remove duplicates
}
