% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_algorithm.R
\name{score_feature}
\alias{score_feature}
\title{Compute Join Aligner Score}
\usage{
score_feature(dM, aM, tolM, drt, aR, tolR)
}
\arguments{
\item{aM}{numeric the mass score contribution (a+b must add up to one)}

\item{aR}{numeric the RT score contribition (see aM)}

\item{m0}{numeric reference mass}

\item{m1}{numeric query mass}

\item{r0}{numeric reference rt}

\item{r1}{numeric query rt}
}
\value{
numeric the feature score
}
\description{
Performs join aligner score to produce the feature score
}
\examples{
score_feature(dM = 1.2,aM = 0.5,tolM = 5, drt = 0.2, aR = 0.5, tolR = 0.3)
0.7936826

score_feature(dM = 0.2,aM = 0.5,tolM = 5, drt = 0.1, aR = 0.5, tolR = 0.3)
0.9533745

}
