% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_nlm.R
\name{search_adduct}
\alias{search_adduct}
\title{Search for adducts using the NL model method}
\usage{
search_adduct(adduct, simdata, ppm_window = 30, instrument_tol = 5,
  alpha_mz = 0.5, rt_tol = 0.2, beta_rt = 0.5)
}
\arguments{
\item{adduct}{numeric value for the adduct search neutral loss mass}

\item{simdata}{list simdata from the features}

\item{instrument_tol}{number for the extpected  mass accuracy of the instrument in ppm}

\item{rt_tol}{numeric tolerance window in minutes or seconds for the dRT score}

\item{tol}{numeric tolerance window in ppm (usually 25-50)}

\item{aMZ}{numeric weighted value for contribution to feature score from the dMZ score}

\item{bRT}{numeric weighted value for contribution to feature score from the dRT score}
}
\value{
list of Search result objects.
}
\description{
Builds a dM model by searching for a specific adduct between MS1 - MS2 datasets.
}
\examples{
search_adduct(-116.0473, simdata)

}
