% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_delta_mass.R
\name{search_delta_mass}
\alias{search_delta_mass}
\title{search_delta_mass}
\usage{
search_delta_mass(type = "SIM", rt_tol = 9, ppm_tol = 20)
}
\arguments{
\item{type}{the search type, either 'SIM' or other for pequdo SIM \(eg. 0-25V consecutive MS2 cycles\).}

\item{rt_tol}{the retention time tolerance to search for the neutral loss mass aglycone peak in seconds.  Usually set to the time for 2 duty cycles.}

\item{ppm_tol}{the mass tolerance in ppm.  Use a wide window, typically 20-50 to aid detection of distribution of mass errors.}
}
\description{
search_delta_mass
}
