% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_nlm.R
\name{search_mass}
\alias{search_mass}
\title{Search masses for adduct neutral loss mass}
\usage{
search_mass(search_list, nldf)
}
\arguments{
\item{nldf}{data.frame the  ms2 data fram or a searchable mass data.frame}

\item{mz}{number the search mz or mass}
}
\value{
class search_result
}
\description{
Search masses for adduct neutral loss mass
}
