% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_NLM.R
\name{search_NLM}
\alias{search_NLM}
\title{search_NLM Wrappper function to search for the neutral loss in wide SIM DIA data.}
\usage{
search_NLM(
  file = NULL,
  type = "SIM",
  raw = FALSE,
  delta_search_mass = NULL,
  search_name = "",
  scandef_file = NULL,
  sim_width = NULL,
  overlapMZ = NULL,
  ppm_tol = 20,
  rt_tol = 9,
  mzmin = 135,
  mzmax = 634,
  mzwid = 0.01
)
}
\arguments{
\item{file}{the mzml file to search. Must be wide SIM MS1 / MS2 DIA data with SIM segments the same mz range}

\item{type}{'SIM' for true wide SIM DIA data or any other value for pseudo MS1 0 volt data \(true MS2\)}

\item{raw}{the Thermo .raw file}

\item{delta_search_mass}{decimal value indicating the Da loss or gain.  Must be signed if negative \(-\)}

\item{search_name}{character value indicating the value used}

\item{scandef_file}{the name of the scan definition file. Currently not used.}

\item{sim_width}{the mz width of the wSIM windows}

\item{overlapMZ}{the mz overlap (if any) between wSIM windows}

\item{ppm_tol}{the ppm mass tolerance in parts per million}

\item{rt_tol}{the retention time tolerance for the search in minutes}

\item{mzmin}{the minimum mz, default = 135}

\item{mzmax}{the maximum mz, default = 634}

\item{mzwid}{the mz slice size, default = 0.01 for Orbitrap}
}
\description{
search_NLM Wrappper function to search for the neutral loss in wide SIM DIA data.
}
