% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_searchResult.R
\docType{data}
\name{search_result}
\alias{search_result}
\title{SearchResult}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
search_result
}
\description{
Defines a SearchResult class  to store search result information.
}
\section{Slots}{

\describe{
\item{\code{index}}{numeric the peak index}

\item{\code{searchmz}}{numeric the search mz value}

\item{\code{searchrt}}{numeric the search rt value}

\item{\code{searchIntensity}}{numeric the search intensity}

\item{\code{resultindex}}{numeric vector containing the indexes of features likely arising from the neutral loss or adduct}

\item{\code{resultrt}}{numeric vector containing the retention times of features likely arising from the neutral loss or adduct}

\item{\code{resultIntensity}}{numeric vector containing the intensities of features likely arising from the neutral loss or adduct}

\item{\code{dRT}}{numeric vector containing the dRT values of features likely arising from the neutral loss or adduct}

\item{\code{resultmz}}{numeric vector containing the mz values of features likely arising from the neutral loss or adduct}

\item{\code{adduct}}{character vector containing the type of adduct searched for}

\item{\code{dM}}{numeric vector containing the dMZ of features likely arising from the neutral loss or adduct}

\item{\code{score}}{numeric vector containing the computed log liklihood score of features likely arising from the neutral loss or adduct}

\item{\code{prob}}{numeric vector containing the maximum liklihood (P(x|u,b) of features likely arising from the neutral loss or adduct}

\item{\code{score_dmz}}{numeric vector containing the dMZ score of features likely arising from the neutral loss or adduct}

\item{\code{score_dRT}}{numeric vector containing the dRT score of features likely arising from the neutral loss or adduct}

\item{\code{score_feat}}{numeric vector containing the total feature score from features likely arising from the neutral loss or adduct}

\item{\code{score_final}}{numeric vector containing the S(feat)x P(feat) of features likely arising from the neutral loss or adduct}
}}

\keyword{datasets}
