% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_convertRAW.R
\name{segmentMzMLData}
\alias{segmentMzMLData}
\title{segmentMzMlData}
\usage{
segmentMzMLData(single_mzml_file, sample_directory_path,
  scandef_file = scandef_file)
}
\arguments{
\item{sample_directory_path}{The path to your outputdata.  The folder must be preexisting.}

\item{mzFile}{The input mzml file}
}
\description{
segmentMzMlData
}
\examples{
#segmentMZMLData("test.mzML", out_path = "results/")
}
