% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_convertRAW.R
\name{segmentMzMLDataSample}
\alias{segmentMzMLDataSample}
\title{Segment mzML data into SIM windows}
\usage{
segmentMzMLDataSample(raw_file_directory, sample_directories,
  scandef_file = scandef_file)
}
\description{
Segregates mzml file into separate mzml files defined by mslevel and sim window.
}
\examples{
#segmentMzMLDataSample

}
