% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_global_opt.R
\name{setGlobal.path}
\alias{setGlobal.path}
\title{setGlobal.path}
\usage{
setGlobal.path(option, val)
}
\arguments{
\item{option}{character vector}

\item{val}{character vector read from the global paramters file.}
}
\description{
Sets the global environment variables called through readGlobalOpts.
}
