% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_scan_def.R
\name{setScanDef}
\alias{setScanDef}
\title{set_scan_def}
\usage{
setScanDef(header, type = "SIM", sim_width = NULL, overlapMZ = NULL)
}
\arguments{
\item{header}{a mzR header object}

\item{type}{either 'SIM' to denote MS1 as SIM spectra or any other as MS1 data collected in the C Trap.}

\item{sim_width}{the emz width of the SIM acquisition window.}

\item{overlapMZ}{optional numeric value describing how much overlap exists between SIM windows.}
}
\value{
dataframe containing scan mz window ranges
}
\description{
set_scan_def
}
