% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_algorithms.R
\name{similarityScore_gauss}
\alias{similarityScore_gauss}
\title{Score similarity using Gaussian function}
\usage{
similarityScore_gauss(deltaVal, tol)
}
\arguments{
\item{deltaVal}{the delta decimal value to score}

\item{tol}{the max decimal tolerance \(eg 0.01 for Da or 3 min for RT\)}
}
\value{
decimal score 0-1
}
\description{
Computes the similarity score using the gaussian assumption.  Use for Da data or RT data.
}
