% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_algorithm.R
\name{similarityScore_laplace}
\alias{similarityScore_laplace}
\title{Score similarity using the Laplace function}
\usage{
similarityScore_laplace(deltaVal, tol)
}
\arguments{
\item{tol}{numeric the tolerance window in generic units}

\item{v0}{numeric the reference value}

\item{v1}{numeric the observed value}
}
\value{
numeric the score
}
\description{
Computes the similarity score usingg the laplace assumption. Used for data measured with ppm errors.
}
\examples{
#' val2 <- 0.21
similarityScore_laplace(val2-val1,tol= 0.1)
0.9048374

val1 <- 0.35
similarityScore_gauss(val2-val1,tol= 0.1)
0.2231302
}
