% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_algorithm.R
\name{weight_gauss}
\alias{weight_gauss}
\title{Apply weights based on mass decimal}
\usage{
weight_gauss(dM, mu, tol, boost)
}
\arguments{
\item{dM}{numeric delta mass}

\item{mu}{numeric mean}

\item{tol}{numeric tolerance}

\item{boost}{numeric integer value to boost signal}
}
\value{
vector of weights
}
\description{
Apply weights based on mass decimal
}
\examples{
weight_gauss(0.1,0.01,5,2)
1.99967
}
