% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/fct_algorithm.R
\name{weight_laplace}
\alias{weight_laplace}
\title{Calculate weight with Laplace assumption.}
\usage{
weight_laplace(dM, mu, tol, boost)

weight_laplace(dM, mu, tol, boost)
}
\arguments{
\item{dM}{vector or numeric the vector of measurement errors}

\item{mu}{numeric the mean of the laplace distribution}

\item{tol}{numeric the ppm tolerance of the instrument}

\item{boost}{numeric the boost value (0-2)}

\item{dM}{vector or numeric the vector of measurement errors}

\item{tol}{numeric the ppm tolerance of the instrument}

\item{boost}{numeric the boost value (0-2)}

\item{mu}{numeric the mean of the laplace distribution}
}
\value{
numeric or vector the weighted value(S).

numeric or vector the weighted value(S).
}
\description{
Assigned a weight to the distribution of X values given expected instrument performance.
For example, dM is the input X values, tol is a tolderance in ppm, and  boost is the multiplier.

Assigned a weight to the distribution of X values given expected instrument performance.
For example, dM is the input X values, tol is a tolderance in ppm, and boost is the multiplier.
}
\examples{
weight_laplace(dM,mu=0.1, tol = 5, boost = 2)
weight_laplace(dM,mu=0.1, tol = 5, boost = 2)
}
