% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_io.R
\name{write_NLM_results}
\alias{write_NLM_results}
\title{Write NLM results}
\usage{
write_NLM_results(searchResultList, fh)
}
\arguments{
\item{searchResultList}{list of class Search objects.}

\item{fh}{character vector file name}
}
\description{
Writes neutral loss search results to file.
}
\examples{
write_NLM_results(searchResultList, fh = "results.tsv")


}
