% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_algorithms.R
\name{getMassTolRange}
\alias{getMassTolRange}
\title{getMassTolRange}
\usage{
getMassTolRange(m, ppm)
}
\arguments{
\item{m}{the mass / mz}

\item{ppm}{the ppm tolerance}
}
\value{
vector of minimum and maximum range
}
\description{
getMassTolRange
}
