% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_nlm.R
\name{getNLMScore}
\alias{getNLMScore}
\title{The the NLM score}
\usage{
getNLMScore(searchResultList, mod_mz)
}
\arguments{
\item{searchResultList}{list of Search class objects}

\item{mod_mz}{list containing the dM mass model results}
}
\value{
list of Search class result objects
}
\description{
Computes the NL model score for each dM value.
}
\examples{

#resultsList <- getNLMScore(searchResultList, mod_mz)


}
