% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_algorithm.R
\name{laplace_unif_EM}
\alias{laplace_unif_EM}
\title{Mixture model: Laplace and Uniform}
\usage{
laplace_unif_EM(x, dM, tol = 1e-05, maxiter = 1000, boost = 2,
  instrument_tol = 0.01)
}
\arguments{
\item{x}{vector of numbers.}

\item{dM}{}

\item{tol}{numeric number indicating tolerance in ppm.}

\item{maxiter}{maximum number of iterations.}

\item{boost}{A weight value applied to Laplace distribution.}

\item{instrument_tol}{}

\item{instrument.tol}{numeric value assigned to}

\item{x}{}

\item{tol}{}

\item{maxiter}{}

\item{boost}{}
}
\value{
list of data model
}
\description{
Performs expectation maximization algorithm on the dM values determined from the data.
}
