% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_convertRAW.R
\name{makeSampleDir}
\alias{makeSampleDir}
\title{Make directories for each sample}
\usage{
makeSampleDir(result_dir, sample_names, scandef_file)
}
\arguments{
\item{result_dir}{character vector pointing to the results directory}

\item{sample_names}{vector of sample names}

\item{scandef_file}{character vector of path to scandef file}
}
\description{
Creates directories for sample of interest indicateg in the global options parameter file.
}
