% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_nlm.R
\name{modelNLM}
\alias{modelNLM}
\title{modelNLM}
\usage{
modelNLM(simdata, neutral_loss_list_file, boost = 2, alpha_mz = 0.5,
  beta_rt = 0.5, ppm_window = 30, rt_tol = 0.2, instrument_tol = 5)
}
\arguments{
\item{simdata}{list of simdata extracted from the feature finding results}

\item{boost}{numeric value for signal boosting during the EM model step}

\item{rt_tol}{numeric}

\item{instrument_tol}{numeric instrument tolerance value in ppm, usually around 5 for an orbitrap.}

\item{fh}{character vector containing the file name for writing results}

\item{aMZ}{numeric weighted value for contribution to feature score from the dMZ score}

\item{bRT}{numeric weighted value for contribution to feature score from the dRT score}

\item{tol}{numeric tolerance window to which perform dM searching.   Needs to be ~5-10 times larger than instrument tolerance}
}
\description{
modelNLM
}
\examples{
##modelNLM(simdata,fh = "model_results.tsv")

}
