% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_nlm.R
\name{modelNLM_run}
\alias{modelNLM_run}
\title{Run NLM model}
\usage{
modelNLM_run(msdial_results, sample_directories, adduct_list, boost = 2,
  alpha_mz = 0.5, beta_rt = 0.5, ppm_window = 30, rt_tol = 0.2,
  instrument_tol = 0.01)
}
\arguments{
\item{msdial_results}{}

\item{boost}{}

\item{instrument_tol}{}
}
\value{
list containing search results
}
\description{
Wrapper function to run modelNLM
}
