% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plotXIC}
\alias{plotXIC}
\title{Plot an extracted ion chromatogram}
\usage{
plotXIC(mz, ppm, rt, rt_tol, ms1_mzml_file, ms2_mzml_file,
  plot_dir = NULL)
}
\arguments{
\item{mz}{numeric containing the mz value}

\item{ppm}{numeric the ppm tolerance window to extract the mz value}

\item{rt}{numeric containing the retention time in seconds}

\item{rt_tol}{numeric containing the retention time range to extract the XIC}

\item{ms1_mzml_file}{character vector listing the ms1 mzML file}

\item{ms2_mzml_file}{character vector listing the ms1 mzML file}

\item{plot_dir}{character vector listing the directory to plot in}
}
\description{
Extracts a smoothed chormatogram for a mz value and retention time.
}
