% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_nlm.R
\name{search_adduct}
\alias{search_adduct}
\title{Search for adducts using the NL model method}
\usage{
search_adduct(adduct_mass = -116.0473, adduct_name = "dR", data_X,
  data_Y, ppm_window = 30, rt_tol = 0.2, alpha_mz = 0.5,
  beta_rt = 0.5, instrument_tol = 0.01)
}
\arguments{
\item{adduct_mass}{}

\item{adduct_name}{}

\item{data_X}{}

\item{data_Y}{}

\item{ppm_window}{}

\item{rt_tol}{}

\item{alpha_mz}{}

\item{beta_rt}{}

\item{instrument_tol}{}
}
\value{
data.frame of search results
}
\description{
Builds a dM model by searching for a specific adduct between MS1 - MS2 datasets.  Alternatively you can 
search within the same data frame or MS level by setting data_X and data_Y equal to each other.
}
