% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_nlm.R
\name{search_mass}
\alias{search_mass}
\title{Search masses for adduct neutral loss mass}
\usage{
search_mass(data_X_row = NULL, data_Y = NULL,
  adduct_mass = -116.0473, adduct_name = "dR", ppm_window = 30,
  rt_tol = 0.3, alpha_mz = 0.5, beta_rt = 0.5,
  instrument_tol = 0.01)
}
\arguments{
\item{data_X_row}{data frame single row of data frame X.}

\item{data_Y}{data frame data to perform search against.}

\item{adduct_mass}{numeric value of the adduct mass to search.  If a loss, then '-' sign must be used.}

\item{adduct_name}{character vector of the adduct name.}

\item{ppm_window}{numeric value for the ppm search window, usually 30-50 ppm.}

\item{rt_tol}{numeric value for the retention time search window.}

\item{alpha_mz}{numeric value, 0-1 to weight the mz search score.}

\item{beta_rt}{numeric value, 0-1 to weight the rt search score. Alpha and beta must add up to 1.}

\item{instrument_tol}{the instrument tolerance window for the search score, usually 10ppm.}
}
\value{
list of search results.
}
\description{
Search masses for adduct neutral loss mass
}
