% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_convertRAW.R
\name{segmentMzMLDataSample}
\alias{segmentMzMLDataSample}
\title{Segment mzML data into SIM windows}
\usage{
segmentMzMLDataSample(raw_file_directory, sample_directories, scandef_file)
}
\arguments{
\item{raw_file_directory}{relative path to the raw files}

\item{sample_directories}{vector of sample directories}

\item{scandef_file}{a non empty character vector giving the wide SIM scan definition file.}
}
\description{
Segregates mzml file into separate mzml files defined by mslevel and sim window.
}
\examples{
#segmentMzMLDataSample

}
