% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_algorithm.R
\name{similarityScore_gauss}
\alias{similarityScore_gauss}
\title{Score similarity using Gaussian function}
\usage{
similarityScore_gauss(deltaVal, tol)
}
\arguments{
\item{tol}{numeric the tolerance window in generic units}

\item{v0}{numeric the reference value}

\item{v1}{numeric the observed value}
}
\value{
numeric the score
}
\description{
Computes the similarity score using the gaussian assumption.  Use for Da data or RT data.
}
\examples{
val1 <- 0.2
val2 <- 0.21
similarityScore_gauss(val2-val1,tol= 0.1)
0.9950125

val1 <- 0.35
similarityScore_gauss(val2-val1,tol= 0.1)
0.3246525


}
