% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_algorithms.R
\name{similarityScore_laplace}
\alias{similarityScore_laplace}
\title{Score similarity using Gaussian function}
\usage{
similarityScore_laplace(deltaVal, tol)
}
\arguments{
\item{deltaVal}{the delta decimal value in ppm to score}

\item{tol}{the max ppm tolerance}
}
\value{
decimal score 0-1
}
\description{
Score similarity using Gaussian function
}
